+++ modules/apps/asset/asset-test/src/testIntegration/java/com/liferay/asset/exportimport/data/handler/test/AssetCategoryPortletDataHandlerTest.java
@@ -12,16 +12,15 @@
  * details.
  */
 
-package com.liferay.asset.lar.test;
+package com.liferay.asset.exportimport.data.handler.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.asset.tags.constants.AssetTagsAdminPortletKeys;
+import com.liferay.asset.categories.admin.web.constants.AssetCategoriesAdminPortletKeys;
 import com.liferay.exportimport.kernel.lar.DataLevel;
 import com.liferay.exportimport.test.util.lar.BasePortletDataHandlerTestCase;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 
-import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.runner.RunWith;
@@ -30,19 +29,13 @@ import org.junit.runner.RunWith;
  * @author Zoltan Csaszi
  */
 @RunWith(Arquillian.class)
-public class AssetTagsPortletDataHandlerTest
+public class AssetCategoryPortletDataHandlerTest
 	extends BasePortletDataHandlerTestCase {
 
 	@ClassRule
 	@Rule
 	public static final AggregateTestRule aggregateTestRule =
 		new LiferayIntegrationTestRule();
-
-	@Before
-	@Override
-	public void setUp() throws Exception {
-		super.setUp();
-	}
 
 	@Override
 	protected void addStagedModels() throws Exception {
@@ -55,7 +48,7 @@ public class AssetTagsPortletDataHandlerTest
 
 	@Override
 	protected String getPortletId() {
-		return AssetTagsAdminPortletKeys.ASSET_TAGS_ADMIN;
+		return AssetCategoriesAdminPortletKeys.ASSET_CATEGORIES_ADMIN;
 	}
 
 	@Override
