+++ modules/apps/asset/asset-test/src/testIntegration/java/com/liferay/asset/exportimport/data/handler/test/AssetTagStagedModelDataHandlerTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.asset.lar.test;
+package com.liferay.asset.exportimport.data.handler.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.asset.kernel.model.AssetTag;
