+++ modules/apps/asset/asset-test/src/testIntegration/java/com/liferay/asset/exportimport/data/handler/test/AssetTagsPortletDataHandlerTest.java
@@ -12,15 +12,16 @@
  * details.
  */
 
-package com.liferay.asset.lar.test;
+package com.liferay.asset.exportimport.data.handler.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.asset.categories.navigation.constants.AssetCategoriesNavigationPortletKeys;
+import com.liferay.asset.tags.constants.AssetTagsAdminPortletKeys;
 import com.liferay.exportimport.kernel.lar.DataLevel;
 import com.liferay.exportimport.test.util.lar.BasePortletDataHandlerTestCase;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 
+import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.runner.RunWith;
@@ -29,7 +30,7 @@ import org.junit.runner.RunWith;
  * @author Zoltan Csaszi
  */
 @RunWith(Arquillian.class)
-public class AssetCategoriesNavigationPortletDataHandlerTest
+public class AssetTagsPortletDataHandlerTest
 	extends BasePortletDataHandlerTestCase {
 
 	@ClassRule
@@ -37,18 +38,24 @@ public class AssetCategoriesNavigationPortletDataHandlerTest
 	public static final AggregateTestRule aggregateTestRule =
 		new LiferayIntegrationTestRule();
 
+	@Before
+	@Override
+	public void setUp() throws Exception {
+		super.setUp();
+	}
+
 	@Override
 	protected void addStagedModels() throws Exception {
 	}
 
 	@Override
 	protected DataLevel getDataLevel() {
-		return DataLevel.PORTLET_INSTANCE;
+		return DataLevel.SITE;
 	}
 
 	@Override
 	protected String getPortletId() {
-		return AssetCategoriesNavigationPortletKeys.ASSET_CATEGORIES_NAVIGATION;
+		return AssetTagsAdminPortletKeys.ASSET_TAGS_ADMIN;
 	}
 
 	@Override
@@ -58,12 +65,12 @@ public class AssetCategoriesNavigationPortletDataHandlerTest
 
 	@Override
 	protected boolean isDataPortletInstanceLevel() {
-		return true;
+		return false;
 	}
 
 	@Override
 	protected boolean isDataSiteLevel() {
-		return false;
+		return true;
 	}
 
 }
