+++ modules/apps/template/template-test/src/testIntegration/java/com/liferay/template/exportimport/data/handler/test/TemplateEntryStagedModelDataHandlerTest.java
@@ -12,33 +12,41 @@
  * details.
  */
 
-package com.liferay.asset.lar.test;
+package com.liferay.template.exportimport.data.handler.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.asset.kernel.model.AssetVocabulary;
-import com.liferay.asset.kernel.service.AssetVocabularyLocalServiceUtil;
-import com.liferay.asset.test.util.AssetTestUtil;
+import com.liferay.dynamic.data.mapping.constants.DDMTemplateConstants;
+import com.liferay.dynamic.data.mapping.model.DDMTemplate;
+import com.liferay.dynamic.data.mapping.service.DDMTemplateLocalService;
 import com.liferay.exportimport.test.util.lar.BaseStagedModelDataHandlerTestCase;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.StagedModel;
+import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.template.TemplateConstants;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
+import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.kernel.util.LocaleUtil;
+import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
+import com.liferay.template.model.TemplateEntry;
+import com.liferay.template.service.TemplateEntryLocalService;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
-import org.junit.Assert;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.runner.RunWith;
 
 /**
- * @author Máté Thurzó
+ * @author Eudaldo Alonso
  */
 @RunWith(Arquillian.class)
-public class AssetVocabularyStagedModelDataHandlerTest
+public class TemplateEntryStagedModelDataHandlerTest
 	extends BaseStagedModelDataHandlerTestCase {
 
 	@ClassRule
@@ -52,45 +60,41 @@ public class AssetVocabularyStagedModelDataHandlerTest
 			Map<String, List<StagedModel>> dependentStagedModelsMap)
 		throws Exception {
 
-		return AssetTestUtil.addVocabulary(group.getGroupId());
+		DDMTemplate ddmTemplate = _ddmTemplateLocalService.addTemplate(
+			TestPropsValues.getUserId(), group.getGroupId(),
+			_portal.getClassNameId(TemplateEntry.class), 0,
+			_portal.getClassNameId(TemplateEntry.class),
+			Collections.singletonMap(LocaleUtil.US, "name"),
+			Collections.emptyMap(), DDMTemplateConstants.TEMPLATE_TYPE_DISPLAY,
+			StringPool.BLANK, TemplateConstants.LANG_TYPE_FTL,
+			"<#-- Empty script -->", new ServiceContext());
+
+		return _templateEntryLocalService.addTemplateEntry(
+			TestPropsValues.getUserId(), group.getGroupId(),
+			ddmTemplate.getTemplateId(), StringPool.BLANK, StringPool.BLANK,
+			new ServiceContext());
 	}
 
 	@Override
 	protected StagedModel getStagedModel(String uuid, Group group)
 		throws PortalException {
 
-		return AssetVocabularyLocalServiceUtil.
-			getAssetVocabularyByUuidAndGroupId(uuid, group.getGroupId());
+		return _templateEntryLocalService.getTemplateEntryByUuidAndGroupId(
+			uuid, group.getGroupId());
 	}
 
 	@Override
 	protected Class<? extends StagedModel> getStagedModelClass() {
-		return AssetVocabulary.class;
+		return TemplateEntry.class;
 	}
 
-	@Override
-	protected void validateImportedStagedModel(
-			StagedModel stagedModel, StagedModel importedStagedModel)
-		throws Exception {
+	@Inject
+	private DDMTemplateLocalService _ddmTemplateLocalService;
 
-		super.validateImportedStagedModel(stagedModel, importedStagedModel);
+	@Inject
+	private Portal _portal;
 
-		AssetVocabulary vocabulary = (AssetVocabulary)stagedModel;
-		AssetVocabulary importedVocabulary =
-			(AssetVocabulary)importedStagedModel;
-
-		Assert.assertEquals(
-			vocabulary.getExternalReferenceCode(),
-			importedVocabulary.getExternalReferenceCode());
-		Assert.assertEquals(vocabulary.getName(), importedVocabulary.getName());
-		Assert.assertEquals(
-			vocabulary.getTitle(LocaleUtil.getDefault()),
-			importedVocabulary.getTitle(LocaleUtil.getDefault()));
-		Assert.assertEquals(
-			vocabulary.getDescription(LocaleUtil.getDefault()),
-			importedVocabulary.getDescription(LocaleUtil.getDefault()));
-		Assert.assertEquals(
-			vocabulary.getSettings(), importedVocabulary.getSettings());
-	}
+	@Inject
+	private TemplateEntryLocalService _templateEntryLocalService;
 
 }
